+++ tests/ignores.txt
@@ -4,3 +4,4 @@ scrapy/interfaces.py
 scrapy/mail.py
 scrapy/utils/console.py
 scrapy/utils/gz.py
+scrapy/utils/job.py
